Basic working version. Needs work still.
